requirements.txt
aiohttp==3.9.1
beautifulsoup4==4.12.2
lxml==4.9.3
google-generativeai>=0.8.0
